requirements.txt
asgiref==3.7.2
Django==4.2.7
django-bootstrap-modal-forms==3.0.4
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-js-asset==2.1.0
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
Pillow==10.1.0
psycopg==3.1.12
psycopg-binary==3.1.12
pytz==2023.3.post1
simplecep==0.1.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
